Remove some inclusion of magic enum
